feat: submit 함수 분리
